Add a module to handle IRC hooks
Doesn't have much for now, but this module is expected to handle _all_
IRC hook handling, and code anywhere else will need to go through this
module to interface with the server. By ensuring that all operations go
through the same common interface, it becomes easier to abstract it away
when we need to test. This won't be for now, though.