Add a hook to fetch and store all IRC server features
This is an improved version of the hook currently in src/wolfgame.py
which doesn't populate the 'var' namespace with cherry-picked features.
Instead, it unconditionally fills in the src.context.Features dict with
every feature the server has available. It's then the responsibility of
other code to access and use the stored information.

The Features dict lives under the src.context namespace because other
modules will need it, and those modules may not be able to import
src.hooks (since the latter might import them). To avoid stupid
workarounds, it is stored in a lower layer.