Update the IRC message handler
This uses the new Features dictionary to properly access and use the IRC
server features. It also removes some weird hacks which are no longer
needed, since we now hold sane defaults in the Features dict.

This also changes the arguments passed to commands; previously an
IRCClient instance would be passed, followed by the nick (or raw nick,
depending on the setting) as a string, then the channel (or the user's
nick, same as previous) as a string, and finally the rest of the message
as a string.

The IRC client has been removed from the parameters, since users and
channels now all hold a reference to it. The nick and channel (which are
better described as the source and the target) are now proper
src.context.IRCContext instances (a user or a channel). The first
argument is now the 'var' namespace, so that external code doesn't need
to import the game state.

The source being the same as the target when a message is sent directly
to the bot is a detail that may also change in the future, and if so,
will probably be replaced by a method or property on the User class.