Make is_admin and is_owner methods of User
Since users are now proper objects of their own, it is much easier to
check whether or not someone is an admin or owner.

Flags may need to be moved over from the global 'var' namespace to each
individual user, but that will be for later.

This also removes a syntax error that slipped through earlier since git
was confused by the line-by-line commit, and a line ended up in the
wrong place.