Add support for WHO requests in hooks
This adds a function hooks.who, which is used to send out WHO requests.
The function properly handles servers with or without WHOX support.

This also adds hooks for the two possible kinds of WHO replies; one for
servers that support WHOX and one for those that don't. This is,
however, completely transparent to the user code, since both hooks fire
the exact same event, and feed it the relevant information.

This also adds a hook for the end of a WHO request, which also fires an
event. For most cases, this is the only event the code will need to hook
into, since the users will be available in the user registry by the time
the WHO responses have been fully processed.