Add a Channel.update_modes method
Since the Channel instances keep track of their own modes (and User
instances keep track of their own status modes in the channels they
occupy), it falls back on the Channel class to update the variables so
that modes are always properly tracked.

This method will be called by the MODE hook, so that the bot always
keeps an accurate registry of channel modes.